Extract audit event creation to separate class
- Added MessagesExporter interface
- Added AuditingMessagesExporter implementation
- Extracted command creation from DecoratingMessagesExporter to allow AuditingMessagesExporter to implement the same interface

CommandFactory has gotten a bit big, but a big advantage is that now all the weird mappings from Search, Query etc. are encapsulated there.
