+++ graylog2-server/src/main/java/org/graylog/plugins/views/ViewsBindings.java
@@ -38,8 +38,10 @@ import org.graylog.plugins.views.search.db.SearchesCleanUpJob;
 import org.graylog.plugins.views.search.elasticsearch.ESGeneratedQueryContext;
 import org.graylog.plugins.views.search.elasticsearch.ElasticsearchQueryString;
 import org.graylog.plugins.views.search.export.ChunkDecorator;
+import org.graylog.plugins.views.search.export.DecoratingMessagesExporter;
 import org.graylog.plugins.views.search.export.ExportBackend;
 import org.graylog.plugins.views.search.export.LegacyChunkDecorator;
+import org.graylog.plugins.views.search.export.MessagesExporter;
 import org.graylog.plugins.views.search.export.SimpleMessageChunkCsvWriter;
 import org.graylog.plugins.views.search.export.es.ElasticsearchExportBackend;
 import org.graylog.plugins.views.search.export.es.RequestStrategy;
@@ -144,6 +146,7 @@ public class ViewsBindings extends ViewsModule {
         bind(ExportBackend.class).to(ElasticsearchExportBackend.class);
         bind(RequestStrategy.class).to(Scroll.class);
         bind(ChunkDecorator.class).to(LegacyChunkDecorator.class);
+        bind(MessagesExporter.class).to(DecoratingMessagesExporter.class);
 
         registerWidgetConfigSubtypes();
