+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/export/CommandFactory.java
@@ -18,6 +18,7 @@ package org.graylog.plugins.views.search.export;
 
 import org.graylog.plugins.views.search.Query;
 import org.graylog.plugins.views.search.Search;
+import org.graylog.plugins.views.search.SearchType;
 import org.graylog.plugins.views.search.elasticsearch.ElasticsearchQueryString;
 import org.graylog.plugins.views.search.searchtypes.MessageList;
 import org.graylog2.plugin.indexer.searches.timeranges.AbsoluteRange;
@@ -49,7 +50,9 @@ public class CommandFactory {
         return builder.build();
     }
 
-    public ExportMessagesCommand buildWithSearchOnly(Search search, Query query, ResultFormat resultFormat) {
+    public ExportMessagesCommand buildWithSearchOnly(Search search, ResultFormat resultFormat) {
+        Query query = queryFrom(search);
+
         return builderFrom(resultFormat)
                 .timeRange(toAbsolute(query.timerange()))
                 .queryString(queryStringFrom(search, query))
@@ -57,7 +60,19 @@ public class CommandFactory {
                 .build();
     }
 
-    public ExportMessagesCommand buildWithMessageList(Search search, Query query, MessageList messageList, ResultFormat resultFormat) {
+    private Query queryFrom(Search s) {
+        if (s.queries().size() > 1) {
+            throw new ExportException("Can't get messages for search with id " + s.id() + ", because it contains multiple queries");
+        }
+
+        return s.queries().stream().findFirst()
+                .orElseThrow(() -> new ExportException("Invalid Search object with empty Query"));
+    }
+
+    public ExportMessagesCommand buildWithMessageList(Search search, String messageListId, ResultFormat resultFormat) {
+        Query query = search.queryForSearchType(messageListId);
+        MessageList messageList = messageListFrom(query, messageListId);
+
         ExportMessagesCommand.Builder commandBuilder = builderFrom(resultFormat)
                 .timeRange(toAbsolute(timeRangeFrom(query, messageList)))
                 .queryString(queryStringFrom(search, query, messageList))
@@ -65,6 +80,17 @@ public class CommandFactory {
                 .decorators(messageList.decorators());
 
         return commandBuilder.build();
+    }
+
+    private MessageList messageListFrom(Query query, String searchTypeId) {
+        SearchType searchType = query.searchTypes().stream()
+                .filter(st -> st.id().equals(searchTypeId))
+                .findFirst().orElseThrow(() -> new IllegalArgumentException("Error getting search type"));
+
+        if (!(searchType instanceof MessageList)) {
+            throw new ExportException("export is not supported for search type " + searchType.getClass());
+        }
+        return (MessageList) searchType;
     }
 
     private AbsoluteRange toAbsolute(TimeRange timeRange) {
