+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/export/CommandFactoryTest.java
@@ -23,6 +23,7 @@ import org.graylog.plugins.views.search.elasticsearch.ElasticsearchQueryString;
 import org.graylog.plugins.views.search.filter.AndFilter;
 import org.graylog.plugins.views.search.filter.StreamFilter;
 import org.graylog.plugins.views.search.searchtypes.MessageList;
+import org.graylog.plugins.views.search.searchtypes.pivot.Pivot;
 import org.graylog2.decorators.Decorator;
 import org.graylog2.plugin.indexer.searches.timeranges.AbsoluteRange;
 import org.graylog2.plugin.indexer.searches.timeranges.InvalidRangeParametersException;
@@ -35,6 +36,7 @@ import java.util.Arrays;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.graylog.plugins.views.search.export.ExportMessagesCommand.DEFAULT_FIELDS;
 import static org.graylog.plugins.views.search.export.ExportMessagesCommand.defaultTimeRange;
 import static org.graylog.plugins.views.search.export.TestData.relativeRange;
@@ -73,6 +75,28 @@ class CommandFactoryTest {
     }
 
     @Test
+    void throwsIfSearchTypeIsNotMessageList() {
+        Pivot p = Pivot.builder().id("pivot-id").series(newArrayList()).rollup(false).build();
+
+        Query q = org.graylog.plugins.views.search.TestData.validQueryBuilder().searchTypes(ImmutableSet.of(p)).build();
+
+        Search s = searchWithQueries(q);
+
+        assertThatExceptionOfType(ExportException.class)
+                .isThrownBy(() -> sut.buildWithMessageList(s, p.id(), ResultFormat.builder().build()))
+                .withMessageContaining("supported");
+    }
+
+    @Test
+    void searchWithMultipleQueriesLeadsToExceptionIfNoSearchTypeProvided() {
+        Search s = searchWithQueries(org.graylog.plugins.views.search.TestData.validQueryBuilder().build(), org.graylog.plugins.views.search.TestData.validQueryBuilder().build());
+
+        assertThatExceptionOfType(ExportException.class)
+                .isThrownBy(() -> sut.buildWithSearchOnly(s, ResultFormat.builder().build()))
+                .withMessageContaining("multiple queries");
+    }
+
+    @Test
     void convertsTimeRangeToAbsolute() {
         RelativeRange relative = relativeRange(100);
         MessagesRequest request = MessagesRequest.builder().timeRange(relative).build();
@@ -90,7 +114,7 @@ class CommandFactoryTest {
                 .build();
         Search s = searchWithQueries(query);
 
-        ExportMessagesCommand command = buildFrom(s, query);
+        ExportMessagesCommand command = buildFrom(s);
 
         assertThat(command.timeRange()).isEqualTo(query.timerange());
         assertThat(command.queryString()).isEqualTo(query.query());
@@ -107,7 +131,7 @@ class CommandFactoryTest {
                 .limit(100)
                 .build();
 
-        ExportMessagesCommand command = buildFrom(s, query, resultFormat);
+        ExportMessagesCommand command = buildFrom(s, resultFormat);
 
         assertThat(command.fieldsInOrder()).isEqualTo(resultFormat.fieldsInOrder());
         assertThat(command.limit()).isEqualTo(resultFormat.limit());
@@ -121,7 +145,7 @@ class CommandFactoryTest {
         ResultFormat resultFormat = ResultFormat.builder()
                 .build();
 
-        ExportMessagesCommand command = buildFrom(s, query, resultFormat);
+        ExportMessagesCommand command = buildFrom(s, resultFormat);
 
         assertThat(command.fieldsInOrder()).isEqualTo(DEFAULT_FIELDS);
     }
@@ -135,7 +159,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.streams()).isEqualTo(ml.effectiveStreams());
     }
@@ -148,7 +172,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.streams()).isEqualTo(q.usedStreamIds());
     }
@@ -162,7 +186,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         //noinspection OptionalGetWithoutIsPresent
         assertThat(command.queryString()).isEqualTo(ml.query().get());
@@ -178,7 +202,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.queryString()).isEqualTo(q.query());
     }
@@ -194,7 +218,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         ElasticsearchQueryString combined = ElasticsearchQueryString.builder()
                 .queryString("from-query AND from-messagelist").build();
@@ -212,7 +236,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.timeRange()).isEqualTo(messageListTimeRange);
     }
@@ -225,7 +249,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.timeRange()).isEqualTo(q.timerange());
     }
@@ -240,7 +264,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml);
+        ExportMessagesCommand command = buildFrom(s, ml.id());
 
         assertThat(command.decorators()).containsExactly(decorator);
     }
@@ -254,7 +278,7 @@ class CommandFactoryTest {
 
         ResultFormat resultFormat = ResultFormat.builder().fieldsInOrder("field-1", "field-2").build();
 
-        ExportMessagesCommand command = buildFrom(s, q, ml, resultFormat);
+        ExportMessagesCommand command = buildFrom(s, ml.id(), resultFormat);
 
         assertThat(command.fieldsInOrder()).isEqualTo(resultFormat.fieldsInOrder());
     }
@@ -267,7 +291,7 @@ class CommandFactoryTest {
 
         Search s = searchWithQueries(q);
 
-        ExportMessagesCommand command = buildFrom(s, q, ml, ResultFormat.builder().build());
+        ExportMessagesCommand command = buildFrom(s, ml.id(), ResultFormat.builder().build());
 
         assertThat(command.fieldsInOrder()).isEqualTo(DEFAULT_FIELDS);
     }
@@ -279,25 +303,25 @@ class CommandFactoryTest {
 
         when(queryStringDecorator.decorateQueryString("undecorated", s, q)).thenReturn("decorated");
 
-        ExportMessagesCommand command = buildFrom(s, q);
+        ExportMessagesCommand command = buildFrom(s);
 
         assertThat(command.queryString()).isEqualTo(ElasticsearchQueryString.builder().queryString("decorated").build());
     }
 
-    private ExportMessagesCommand buildFrom(Search s, Query query) {
-        return buildFrom(s, query, ResultFormat.builder().build());
+    private ExportMessagesCommand buildFrom(Search s) {
+        return buildFrom(s, ResultFormat.builder().build());
     }
 
-    private ExportMessagesCommand buildFrom(Search s, Query query, ResultFormat resultFormat) {
-        return sut.buildWithSearchOnly(s, query, resultFormat);
+    private ExportMessagesCommand buildFrom(Search s, ResultFormat resultFormat) {
+        return sut.buildWithSearchOnly(s, resultFormat);
     }
 
-    private ExportMessagesCommand buildFrom(Search search, Query query, MessageList messageList) {
-        return buildFrom(search, query, messageList, ResultFormat.builder().build());
+    private ExportMessagesCommand buildFrom(Search search, String messageListId) {
+        return buildFrom(search, messageListId, ResultFormat.builder().build());
     }
 
-    private ExportMessagesCommand buildFrom(Search search, Query query, MessageList messageList, ResultFormat resultFormat) {
-        return sut.buildWithMessageList(search, query, messageList, resultFormat);
+    private ExportMessagesCommand buildFrom(Search search, String messageListId, ResultFormat resultFormat) {
+        return sut.buildWithMessageList(search, messageListId, resultFormat);
     }
 
     private AndFilter streamFilter(String... streamIds) {
