+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/rest/MessagesResourceTest.java
@@ -18,9 +18,12 @@ package org.graylog.plugins.views.search.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.eventbus.EventBus;
 import org.graylog.plugins.views.search.SearchDomain;
 import org.graylog.plugins.views.search.SearchExecutionGuard;
 import org.graylog.plugins.views.search.errors.PermissionException;
+import org.graylog.plugins.views.search.export.CommandFactory;
+import org.graylog.plugins.views.search.export.ExportMessagesCommand;
 import org.graylog.plugins.views.search.export.MessagesExporter;
 import org.graylog.plugins.views.search.export.MessagesRequest;
 import org.graylog2.plugin.database.users.User;
@@ -31,12 +34,12 @@ import org.mockito.ArgumentCaptor;
 
 import javax.annotation.Nullable;
 import java.util.Collections;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -48,6 +51,9 @@ public class MessagesResourceTest {
     private MessagesExporter exporter;
     private PermittedStreams permittedStreams;
     private SearchExecutionGuard executionGuard;
+    private CommandFactory commandFactory;
+    @SuppressWarnings("UnstableApiUsage")
+    private final EventBus eventBus = mock(EventBus.class);
 
     @BeforeEach
     void setUp() {
@@ -55,10 +61,14 @@ public class MessagesResourceTest {
         currentUser = mock(User.class);
         when(currentUser.getName()).thenReturn("peterchen");
         exporter = mock(MessagesExporter.class);
+        commandFactory = mock(CommandFactory.class);
+        when(commandFactory.buildFromRequest(any())).thenReturn(ExportMessagesCommand.withDefaults());
+        when(commandFactory.buildWithSearchOnly(any(), any())).thenReturn(ExportMessagesCommand.withDefaults());
+        when(commandFactory.buildWithMessageList(any(), any(), any())).thenReturn(ExportMessagesCommand.withDefaults());
         permittedStreams = mock(PermittedStreams.class);
         when(permittedStreams.load(any())).thenReturn(ImmutableSet.of("a-default-stream"));
         executionGuard = mock(SearchExecutionGuard.class);
-        sut = new MessagesTestResource(exporter, mock(SearchDomain.class), executionGuard, permittedStreams, mock(ObjectMapper.class));
+        sut = new MessagesTestResource(exporter, commandFactory, mock(SearchDomain.class), executionGuard, permittedStreams, mock(ObjectMapper.class), eventBus);
 
         sut.asyncRunner = c -> {
             c.accept(x -> {
@@ -68,8 +78,8 @@ public class MessagesResourceTest {
     }
 
     class MessagesTestResource extends MessagesResource {
-        public MessagesTestResource(MessagesExporter exporter, SearchDomain searchDomain, SearchExecutionGuard executionGuard, PermittedStreams permittedStreams, ObjectMapper objectMapper) {
-            super(exporter, searchDomain, executionGuard, permittedStreams, objectMapper);
+        public MessagesTestResource(MessagesExporter exporter, CommandFactory commandFactory, SearchDomain searchDomain, SearchExecutionGuard executionGuard, PermittedStreams permittedStreams, ObjectMapper objectMapper, EventBus eventBus) {
+            super(exporter, commandFactory, searchDomain, executionGuard, permittedStreams, objectMapper, eventBus);
         }
 
         @Nullable
@@ -87,11 +97,12 @@ public class MessagesResourceTest {
 
         ArgumentCaptor<MessagesRequest> captor = ArgumentCaptor.forClass(MessagesRequest.class);
 
-        doNothing().when(exporter).export(captor.capture(), any(), any());
+        when(commandFactory.buildFromRequest(captor.capture())).thenReturn(ExportMessagesCommand.withDefaults());
 
         sut.retrieve(request);
 
-        assertThat(captor.getValue().streams())
+        MessagesRequest value = captor.getValue();
+        assertThat(value.streams())
                 .containsExactly("stream-1", "stream-2");
     }
 
@@ -105,6 +116,19 @@ public class MessagesResourceTest {
 
         assertThatExceptionOfType(PermissionException.class).isThrownBy(() -> sut.retrieve(request))
                 .withMessageContaining(exception.getMessage());
+    }
+
+    @Test
+    void passesUserNameToAuditingExporter() {
+        AtomicReference<String> passedUsername = new AtomicReference<>();
+        sut.messagesExporterFactory = userName -> {
+            passedUsername.set(userName);
+            return mock(MessagesExporter.class);
+        };
+
+        sut.retrieve(validRequest());
+
+        assertThat(passedUsername.toString()).isEqualTo(currentUser.getName());
     }
 
     private MessagesRequest validRequest() {
